requirements.txt
ultralytics
opencv-python
numpy
pyyaml
supervision
pandas
matplotlib
seaborn
